+++ examples/simple/bug-1149658.c
@@ -34,7 +34,7 @@ int main() {
                                 /*membership=*/ 0, /*memberships=*/ 0, &mod);
 
     if (igraph_vector_size(&mod) != 1 ||
-        VECTOR(mod)[0] != 0) {
+        !igraph_is_nan(VECTOR(mod)[0])) {
         return 1;
     }